Fixed build errors
