--- samples/csharp/getting-started/CustomVisionObjectDetectionOnnx/CustomVisionObjectDetectionOnnx/DataStructures/ImageNetPrediction.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.ML.Data;
-
-namespace CustomVisionObjectDetectionOnnx.DataStructures
-{
-    public class ImageNetPrediction
-    {
-        [ColumnName("model_outputs0")]
-        public float[] PredictedLabels;
-    }
-}
