+++ samples/csharp/getting-started/DeepLearning_ObjectDetection_Custom_Onnx/CustomObjDetectionOnnx.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.757
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomObjDetectionOnnx", "CustomObjDetectionOnnx\CustomObjDetectionOnnx.csproj", "{F87FD74A-6FFE-4503-B351-4D2805D25284}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F87FD74A-6FFE-4503-B351-4D2805D25284}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F87FD74A-6FFE-4503-B351-4D2805D25284}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F87FD74A-6FFE-4503-B351-4D2805D25284}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F87FD74A-6FFE-4503-B351-4D2805D25284}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {69650D93-94C5-4357-A49F-CFD18B9A7827}
+	EndGlobalSection
+EndGlobal
