+++ samples/csharp/getting-started/DeepLearning_ObjectDetection_Custom_Onnx/CustomObjDetectionOnnx/DataStructures/ImageNetPrediction.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.ML.Data;
+
+namespace CustomVisionObjectDetectionOnnx.DataStructures
+{
+    public class ImageNetPrediction
+    {
+        [ColumnName("model_outputs0")]
+        public float[] PredictedLabels;
+    }
+}
