+++ samples/csharp/getting-started/DeepLearning_ObjectDetection_Custom_Onnx/CustomObjDetectionOnnx/OnnxModelScorer.cs
@@ -1,0 +1,87 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.ML;
+using Microsoft.ML.Data;
+using CustomVisionObjectDetectionOnnx.DataStructures;
+using CustomVisionObjectDetectionOnnx.YoloParser;
+
+namespace CustomVisionObjectDetectionOnnx
+{
+    class OnnxModelScorer
+    {
+        private readonly string imagesFolder;
+        private readonly string modelLocation;
+        private readonly MLContext mlContext;
+
+        private IList<YoloBoundingBox> _boundingBoxes = new List<YoloBoundingBox>();
+
+        public OnnxModelScorer(string imagesFolder, string modelLocation, MLContext mlContext)
+        {
+            this.imagesFolder = imagesFolder;
+            this.modelLocation = modelLocation;
+            this.mlContext = mlContext;
+        }
+
+        public struct ImageNetSettings
+        {
+            public const int imageHeight = 416;
+            public const int imageWidth = 416;
+        }
+
+        public struct TinyYoloModelSettings
+        {
+            // for checking Tiny yolo2 Model input and  output  parameter names,
+            //you can use tools like Netron, 
+            // which is installed by Visual Studio AI Tools
+
+            // input tensor name
+            public const string ModelInput = "data";
+
+            // output tensor name
+            public const string ModelOutput = "model_outputs0";
+        }
+
+        private ITransformer LoadModel(string modelLocation)
+        {
+            Console.WriteLine("Read model");
+            Console.WriteLine($"Model location: {modelLocation}");
+            Console.WriteLine($"Default parameters: image size=({ImageNetSettings.imageWidth},{ImageNetSettings.imageHeight})");
+
+            // Create IDataView from empty list to obtain input data schema
+            var data = mlContext.Data.LoadFromEnumerable(new List<ImageNetData>());
+
+            // Define scoring pipeline
+            var pipeline = mlContext.Transforms.LoadImages(outputColumnName: "data", imageFolder: "", inputColumnName: nameof(ImageNetData.ImagePath))
+                            .Append(mlContext.Transforms.ResizeImages(outputColumnName: "data", imageWidth: ImageNetSettings.imageWidth, imageHeight: ImageNetSettings.imageHeight, inputColumnName: "data"))
+                            .Append(mlContext.Transforms.ExtractPixels(outputColumnName: "data"))
+                            .Append(mlContext.Transforms.ApplyOnnxModel(modelFile: modelLocation, outputColumnNames: new[] { TinyYoloModelSettings.ModelOutput }, inputColumnNames: new[] { TinyYoloModelSettings.ModelInput }));
+
+            // Fit scoring pipeline
+            var model = pipeline.Fit(data);
+
+            return model;
+        }
+
+        private IEnumerable<float[]> PredictDataUsingModel(IDataView testData, ITransformer model)
+        {
+            Console.WriteLine($"Images location: {imagesFolder}");
+            Console.WriteLine("");
+            Console.WriteLine("=====Identify the objects in the images=====");
+            Console.WriteLine("");
+
+            IDataView scoredData = model.Transform(testData);
+
+            IEnumerable<float[]> probabilities = scoredData.GetColumn<float[]>(TinyYoloModelSettings.ModelOutput);
+
+            return probabilities;
+        }
+
+        public IEnumerable<float[]> Score(IDataView data)
+        {
+            var model = LoadModel(modelLocation);
+
+            return PredictDataUsingModel(data, model);
+        }
+    }
+}
