--- samples/csharp/getting-started/DeepLearning_ObjectDetection_Custom_Onnx/CustomVisionObjectDetectionOnnx/DataStructures/ImageNetData.cs
@@ -1,24 +1,0 @@
-﻿using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using Microsoft.ML.Data;
-
-namespace CustomVisionObjectDetectionOnnx.DataStructures
-{
-    public class ImageNetData
-    {
-        [LoadColumn(0)]
-        public string ImagePath;
-
-        [LoadColumn(1)]
-        public string Label;
-
-        public static IEnumerable<ImageNetData> ReadFromFile(string imageFolder)
-        {
-            return Directory
-                .GetFiles(imageFolder)
-                .Where(filePath => Path.GetExtension(filePath) != ".md")
-                .Select(filePath => new ImageNetData { ImagePath = filePath, Label = Path.GetFileName(filePath) });
-        }
-    }
-}
