--- samples/csharp/getting-started/DeepLearning_ObjectDetection_Custom_Onnx/CustomVisionObjectDetectionOnnx/YoloParser/DimensionsBase.cs
@@ -1,10 +1,0 @@
-﻿namespace CustomVisionObjectDetectionOnnx.YoloParser
-{
-    public class DimensionsBase
-    {
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Height { get; set; }
-        public float Width { get; set; }
-    }
-}
