--- samples/csharp/getting-started/DeepLearning_ObjectDetection_Custom_Onnx/CustomVisionObjectDetectionOnnx/YoloParser/YoloBoundingBox.cs
@@ -1,22 +1,0 @@
-﻿using System.Drawing;
-
-namespace CustomVisionObjectDetectionOnnx.YoloParser
-{
-    public class BoundingBoxDimensions : DimensionsBase { }
-
-    public class YoloBoundingBox
-    {
-        public BoundingBoxDimensions Dimensions { get; set; }
-
-        public string Label { get; set; }
-
-        public float Confidence { get; set; }
-
-        public RectangleF Rect
-        {
-            get { return new RectangleF(Dimensions.X, Dimensions.Y, Dimensions.Width, Dimensions.Height); }
-        }
-
-        public Color BoxColor { get; set; }
-    }
-}
